# Conflicts:
#	app/src/main/java/com/tosslab/jandi/app/ui/message/v2/MessageListV2Fragment.java

+++ app/src/main/java/com/tosslab/jandi/app/events/push/MessagePushEvent.java
@@ -4,15 +4,15 @@ package com.tosslab.jandi.app.events.push;
  * Created by Steve SeongUg Jung on 15. 1. 20..
  */
 public class MessagePushEvent {
-    private final int entityId;
+    private final long entityId;
     private final String entityType;
 
-    public MessagePushEvent(int entityId, String entityType) {
+    public MessagePushEvent(long entityId, String entityType) {
         this.entityId = entityId;
         this.entityType = entityType;
     }
 
-    public int getEntityId() {
+    public long getEntityId() {
         return entityId;
     }
